move findMatchingPattern to DeepLinkParser.kt and make it private
